requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
email-validator==2.3.0
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
python-multipart==0.0.6
